# src/backend/routes/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Drivers Logbook Web Application API",
    "version": "1.0.0",
    "description": "API documentation for the Drivers Logbook Web Application."
  },
  "servers": [
    {
      "url": "http://localhost/",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/api/users": {
      "get": {
        "summary": "Retrieves all users.",
        "description": "Returns a list of all users. Requires admin API key.",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "A list of users.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error."
          }
        }
      }
    },
    "/api/user/session": {
      "get": {
        "summary": "Retrieves the session user.",
        "description": "Returns the currently authenticated user's session data.",
        "tags": [
          "Users"
        ],
        "responses": {
          "200": {
            "description": "Session user information.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "type": "object",
                      "properties": {
                        "firstName": {
                          "type": "string"
                        },
                        "lastName": {
                          "type": "string"
                        },
                        "webUserId": {
                          "type": "string"
                        }
                      }
                    },
                    "session": {
                      "type": "object",
                      "properties": {
                        "expireTimestamp": {
                          "type": "number"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized User-Session."
          }
        }
      }
    },
    "/api/user/{id}": {
      "get": {
        "summary": "Retrieves a user by ID.",
        "description": "Returns a specific user based on their ID. Requires admin API key.",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The user ID."
          }
        ],
        "responses": {
          "200": {
            "description": "A user object.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "404": {
            "description": "User not found."
          }
        }
      },
      "patch": {
        "summary": "Updates a user.",
        "description": "Updates user data. The authenticated user can update their own data.",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The user ID to update."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "role": {
                    "type": "string"
                  },
                  "vehicleId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Bad request."
          },
          "401": {
            "description": "Unauthorized."
          }
        }
      },
      "delete": {
        "summary": "Deletes a user.",
        "description": "Deletes a user by ID. Requires admin API key.",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The user ID to delete."
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted successfully."
          },
          "500": {
            "description": "Server error."
          }
        }
      }
    },
    "/api/user": {
      "post": {
        "summary": "Creates a new user.",
        "description": "Creates a new user. Requires admin API key.",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  },
                  "firstName": {
                    "type": "string"
                  },
                  "lastName": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "role": {
                    "type": "string"
                  },
                  "vehicleId": {
                    "type": "string"
                  }
                },
                "required": [
                  "username",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Bad request."
          }
        }
      }
    },
    "/api/login": {
      "post": {
        "summary": "Authenticates a user.",
        "description": "Logs in a user using username and password. On success, a session is created.",
        "tags": [
          "Authentication"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": [
                  "username",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "302": {
            "description": "Redirects to home on successful login."
          },
          "400": {
            "description": "Missing username or password."
          },
          "401": {
            "description": "Invalid username or password."
          },
          "500": {
            "description": "Server error."
          }
        }
      }
    },
    "/api/logout": {
      "post": {
        "summary": "Logs out the current user.",
        "description": "Destroys the user session and clears the session cookie.",
        "tags": [
          "Authentication"
        ],
        "responses": {
          "200": {
            "description": "Logged out successfully."
          },
          "500": {
            "description": "Logout failed."
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "vehicleId": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}